Avoid updating wr timers during sync
The write timers were updated during sync and flush without
checking if the file was opened for a write. This can add
small (negligible) amounts to the write timers when a
file is opened for reading.

We now update the write timers during sync/flush only if the file
was opened for writing